Reorganize crate as library / app
Needed for API testing.
